==> EqSimWHM/Scripts/03_EqSim_summarize.Rmd <==
---
output: 
  word_document:
    reference_docx: ..\..\PFA_report_template_v1.4.dotx
---

```{r setup, include=FALSE}

################################################################################
# EqSim Summarize.Rmd
#
# 06/07/2020 tested on 1000 iters of SAM assessment
# 05/08/2020 converted to Rmd file
#
################################################################################

require("knitr")
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, crop = TRUE, comment = "")
knitr::opts_chunk$set(fig.width=10) 

options(dplyr.summarise.inform = FALSE)

# rm(list=ls())
invisible(gc())

library(FLCore)
library(tidyverse)

Base.dir <- dirname(dirname(getwd()))   # go up two folders from the scripts directory
MSE.dir <- file.path(Base.dir,"EqSimWHM")
Res.dir <- file.path(MSE.dir, "Results")
stats.dir <- file.path(Res.dir, "Stats")

Source.dir <- file.path(MSE.dir,"R")              #R functions
source(file.path(Source.dir,"utilities.R"))

# Rebuilding threshold
rebuiltThreshold <- 0.5

file.list <- list.files(path=stats.dir, pattern="_Stats", full.names=TRUE)

# print(file.list)
i <- 1

df <- data.frame(stringsAsFactors = FALSE)
for (i in 1:length(file.list)) {
  print(file.list[i])
  df <- bind_rows(df, loadRData(file.list[i])[["df"]])
}

# manual fixing of units
my_units <- data.frame(
  perfstat = c("stock","fbar","catch","rec",     "tac", "pblim", "precbpa"),
  unit2     = c(   "kt","F(1-10)",   "kt","billions", "kt", "proportion","proportion"),
  stringsAsFactors = FALSE)

# COULD BE DONE IN SIMULATION CODE
df <- 
  df %>%  
  mutate(method = ifelse(is.na(method),"EQSIM", method)) %>% 
  mutate(period = ifelse(is.na(period) & year <= (an(assessyear)), "HI",period)) %>% 
  mutate(ftgt = ac(ftgt)) %>% 
  
  left_join(my_units, by="perfstat") %>% 
  
  mutate(
    mean = ifelse(perfstat == "stock", mean/1000, mean),
    median = ifelse(perfstat == "stock", median/1000, median),
    lower  = ifelse(perfstat == "stock", lower/1000, lower),
    upper  = ifelse(perfstat == "stock", upper/1000, upper),
    data   = ifelse(perfstat == "stock", data/1000, data),
    blim   = ifelse(perfstat == "stock", blim/1000, blim),
    bpa    = ifelse(perfstat == "stock", bpa/1000, bpa),
    msybtrig  = ifelse(perfstat == "stock", msybtrig/1000, msybtrig),
  ) %>% 
  
  mutate(
    mean = ifelse(perfstat == "rec", mean/1000000, mean),
    median = ifelse(perfstat == "rec", median/1000000, median),
    lower  = ifelse(perfstat == "rec", lower/1000000, lower),
    upper  = ifelse(perfstat == "rec", upper/1000000, upper),
    data   = ifelse(perfstat == "rec", data/1000000, data)
  ) %>% 

  mutate(
    mean = ifelse(perfstat == "catch", mean/1000, mean),
    median = ifelse(perfstat == "catch", median/1000, median),
    lower  = ifelse(perfstat == "catch", lower/1000, lower),
    upper  = ifelse(perfstat == "catch", upper/1000, upper),
    data   = ifelse(perfstat == "catch", data/1000, data)
  ) 

# save(df, file="df.RData")

# plot function
plotvar <- function(mystock      =  "WHOM",
                    myassess     = "SAM",
                    myassessyear = c("2019","2020"),
                    mymethod  = "EQSIM",
                    myom         = c("OM2.4","OM2.5"),
                    myniters     = "1000",
                    mymp         = c("MP5.23"),
                    mynyrs       = "23",
                    myftgt       = ac(c(0.0, 0.025, 0.05, 0.075, 0.10, 0.125, 0.15)),
                    myperfstat   = "stock",
                    mycolour     = "blue",
                    myyintercept = "blim",
                    myvalue      = "median",
                    myfacets     = c("assessyear","ftgt"),
                    myfirstyear  = as.numeric(NA),
                    mylastyear   = as.numeric(NA)) {
  
  
  # history
  h <-
    df %>% 
    filter(period     %in% c("HI")) %>% 
    filter(perfstat   %in% myperfstat) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    # filter(nyrs       %in% mynyrs) %>% 
    filter(ftgt       %in% ac(myftgt)) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    mutate(age = as.numeric(age)) 

  # future
  d <- 
    df %>%
    filter(period     %in% c("CU","ST","MT",'LT')) %>% 
    filter(perfstat   %in% myperfstat) %>% 
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>%
    # filter(nyrs       %in% mynyrs) %>% 
    filter(ftgt       %in% ac(myftgt)) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    mutate(age  = as.numeric(age)) 
    
  # periods
  p <-
    bind_rows(h, d) %>% 
    distinct(assessyear, period, year) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    group_by(assessyear, period) %>% 
    summarise(
      minyear = min(year),
      maxyear = max(year)
    )
  
  # intercept
  if(!all(is.na(myyintercept))) {
    i <-
      h %>% 
      distinct(get(myyintercept)) %>% 
      an()
  } else {
    i <- an(NA)
  }

  # title
  t <- paste(
    toupper(myperfstat),
    toupper(mystock),
    paste(unique(myassess), collapse="-"),
    paste(unique(myassessyear), collapse="-"),
    paste(unique(mymethod), collapse="-"),
    paste(unique(myom), collapse="-"),
    paste(unique(mymp), collapse="-"),
    sep="_"
  )
  
  # h %>% filter(!is.na(iter), !is.na(data)) %>% View()
  
  myfig <- 

    ggplot(d, aes(x=year, y=get(myvalue), group=mp)) +
    theme_publication() +
    theme(axis.text.x = element_text(angle = 90, vjust=0.5)) +
    theme( panel.grid.major.x = element_blank()) +
    theme(legend.position = "none") +
    theme(panel.spacing = unit(0.2, "lines")) +
    
    # historical ribbon
    geom_ribbon(data=filter(h, !is.na(get(myvalue))),
                aes(ymin=lower, ymax=upper, group=iter), alpha=0.4, fill="gray") +
    
    # historical worms
    geom_line(data=filter(h, metric=="worm"),
              aes(x=year, y=data, group=iter), colour="darkgray", size=0.5) +
    
    # historical median
    geom_line(data=filter(h, !is.na(get(myvalue))), 
              aes(x=year,y=get(myvalue),group=iter), colour="black", size=1) +
    
    # future ribbon
    geom_ribbon(data=filter(d, !is.na(get(myvalue))),
                aes(ymin=lower, ymax=upper, group=iter), alpha=0.4, fill=mycolour) +
    
    # future worms
    geom_line(data=filter(d, metric=="worm"),
              aes(x=year, y=data, group=iter), colour=mycolour, size=0.5) +
    
    # future median
    geom_line(data=filter(d, !is.na(get(myvalue))),
              aes(x=year,y=get(myvalue),group=iter), colour=mycolour, size=1) +
    
    # hline
    {if(!all(is.na(myyintercept))) geom_hline(yintercept=i, linetype="dashed")}  +
    
    # vertical lines (periods)
    geom_vline(data=filter(p, period=="CU"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="ST"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="MT"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="LT"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="LT"),
               aes(xintercept=maxyear), linetype="dotted") +
    
    expand_limits(y=0) +
    labs(x="", y=myperfstat, title=t) +
    facet_grid(get(myfacets[1]) ~ get(myfacets[2])) 
    
  print(myfig)
  
  ggsave(file = file.path(Res.dir,paste0(t, " summaryplot.png")),
         device="png", width = 30, height = 20, units = "cm")
  
}

# ========================================================================================================

plotrecovery <- function(
                    mystock      =  "WHOM",
                    myassess     = "SAM",
                    myassessyear = c("2019","2020"),
                    mymethod     = "EQSIM",
                    myom         = c("OM2.4","OM2.5"),
                    mymp         = c("MP5.23"),
                    myftgt       = c(0.0, 0.025, 0.05, 0.075, 0.10, 0.125, 0.15),
                    myyintercept = 0.5,
                    myfacets     = c("assessyear","ftgt"),
                    myfirstyear  = as.numeric(NA),
                    mylastyear   = as.numeric(NA)) {
  
  # plot recovery to blim and bpa
  d <-
    df %>%
    filter(perfstat %in% c("precblim","precbpa")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} 
  
  # lines
  p <-
    df %>%
    filter(perfstat %in% c("firstyearrebuildtoblim","firstyearrebuildtobpa")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    mutate(perfstat = case_when(
      perfstat == "firstyearrebuildtoblim" ~ "precblim",
      perfstat == "firstyearrebuildtobpa"  ~ "precbpa",
      TRUE                                 ~ perfstat
    ))

  # title
  t <- paste(
    "RECOV",
    toupper(mystock),
    paste(unique(myassess), collapse="-"),
    paste(unique(myassessyear), collapse="-"),
    paste(unique(mymethod), collapse="-"),
    paste(unique(myom), collapse="-"),
    paste(unique(mymp), collapse="-"),
    sep="_"
  )
  
  myfig <-
    d %>% 
    ggplot(aes(x=year, y=mean, group=perfstat)) +
    theme_publication() +
    theme(axis.text.x = element_text(angle = 90, vjust=0.5)) +
    theme( panel.grid.major.x = element_blank()) +
    
    geom_vline(data=p, aes(xintercept=mean, colour=perfstat), linetype="dashed") +
    
    geom_hline(yintercept=myyintercept, colour="gray", linetype="dashed") +
    
    geom_text(data=filter(p, perfstat %in% c("precblim")),
              aes(x=mean, label=ac(mean), colour=perfstat),
              y=1.0, vjust=1, hjust=1, nudge_x = -1, size=2) +
    
    geom_text(data=filter(p, perfstat %in% c("precbpa")),
              aes(x=mean, label=ac(mean), colour=perfstat),
              y=0.0, vjust=0, hjust=0, nudge_x = 1, size=2) +
    
    geom_line(aes(colour=perfstat)) +
    
    expand_limits(y=0) +
    labs(x="", y="", title=t) +
    facet_grid(get(myfacets[1]) ~ get(myfacets[2])) 
  
  print(myfig)
  
  ggsave(file = file.path(Res.dir,paste0(t, " recoveryplot.png")),
         device="png", width = 30, height = 20, units = "cm")
  
  
} # end of plotrecovery

# mystock      = "WHOM"
# myassess     = "SS3"
# myassessyear = c("2019")
# mymethod     = c("EQSIM")
# myom         = c("OM2.2")
# mymp         = c("MP5.00","MP5.01","MP5.03")
# myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15)
# myfacets     = c("mp","ftgt")
# myfirstyear  = 2000

plotpblim <- function(
                    mystock      =  "WHOM",
                    myassess     = "SAM",
                    myassessyear = c("2019","2020"),
                    mymethod     = "EQSIM",
                    myom         = c("OM2.4","OM2.5"),
                    mymp         = c("MP5.23"),
                    myftgt       = c(0.0, 0.025, 0.05, 0.075, 0.10, 0.125, 0.15),
                    myfacets     = c("assessyear","ftgt"),
                    myfirstyear  = as.numeric(NA),
                    mylastyear   = as.numeric(NA) ) {
  
  # plot probability below blim and bpa
  d <-
    df %>%
    filter(perfstat   %in% c("pblim")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} 
  
  # periods
  p <-
    d %>% 
    distinct(assessyear, period, year) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    group_by(assessyear, period) %>% 
    summarise(
      minyear = min(year),
      maxyear = max(year)
    )
  
  # title
  t <- paste(
    "pBlim",
    toupper(mystock),
    paste(unique(myassess), collapse="-"),
    paste(unique(myassessyear), collapse="-"),
    paste(unique(mymethod), collapse="-"),
    paste(unique(myom), collapse="-"),
    paste(unique(mymp), collapse="-"),
    sep="_"
  )
  
  myfig <-
    d %>% 
    ggplot(aes(x=year, y=mean, group=perfstat)) +
    theme_publication() +
    theme(axis.text.x = element_text(angle = 90, vjust=0.5)) +
    theme( panel.grid.major.x = element_blank()) +
    
    geom_hline(yintercept=0.05, colour="gray", linetype="dashed") +
    
    geom_line(aes(colour=perfstat)) +
    
    # vertical lines (periods)
    geom_vline(data=filter(p, period=="CU"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="ST"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="MT"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="LT"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="LT"),
               aes(xintercept=maxyear), linetype="dotted") +

    expand_limits(y=0) +
    labs(x="", y="", title=t) +
    facet_grid(get(myfacets[[1]]) ~ get(myfacets[[2]])) 
  
  print(myfig)
  
  ggsave(file = file.path(Res.dir,paste0(t, " pblimplot.png")),
         device="png", width = 30, height = 20, units = "cm")
  
  
} # end of plotpblim




# mystock      =  "WHOM"
# myassess     = "SS3"
# myassessyear = c("2020")
# mymethod     = "EQSIM"
# myom         = c("OM2.3")
# mymp         = c("MP5.23")
# myftgt       = c(0.075)
# myfacets     = c("perfstat")
# myfirstyear  = 2000
# mylastyear   = as.numeric(NA)
# myperfstat   = c("stock","fbar", "pblim", "precbpa")
# mycolours    = c("red","blue","green", "blue")

# mystock      =  "WHOM"
# myassess     = "SS3"
# myassessyear = c("2020")
# mymethod     = "EQSIM"
# myom         = c("OM2.3")
# mymp         = c("MP5.23")
# myftgt       = c("0.074")
# myfacets     = c("perfstat")
# myfirstyear  = 2000
# mylastyear   = as.numeric(NA)
# myperfstat   = c("fbar")
# mycolours    = c("blue")
# mylastyear   = as.numeric(NA)

plotscenario <- function(
                    mystock      =  "WHOM",
                    myassess     = "SS3",
                    myassessyear = c("2020"),
                    mymethod     = "EQSIM",
                    myom         = c("OM2.3"),
                    mymp         = c("MP5.23"),
                    myftgt       = c(0.075),
                    myfacets     = c("perfstat"),
                    myfirstyear  = as.numeric(NA),
                    mylastyear   = as.numeric(NA),
                    myperfstat   = c("stock","pblim"),
                    mycolours    = rep(c("red","blue","green", "blue"),5),
                    showtitle    = FALSE) {
  
  # history
  h <-
    df %>% 
    filter(period     %in% c("HI")) %>% 
    filter(perfstat   %in% myperfstat) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>%
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>% 
    
    mutate(age = as.numeric(age)) 

  # future
  d <- 
    df %>%
    filter(period     %in% c("CU","ST","MT",'LT')) %>% 
    filter(perfstat   %in% myperfstat) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>% 
    
    mutate(age  = as.numeric(age)) 
    
  # periods
  p <-
    bind_rows(h, d) %>% 
    distinct(assessyear, period, year) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    group_by(assessyear, period) %>% 
    summarise(
      minyear = min(year),
      maxyear = max(year)
    )

  # units
  u <-
    d %>% 
    distinct(perfstat, unit2)
  
  # recoverylines
  p2 <-
    df %>%
    filter(perfstat %in% c("firstyearrebuildtobpa")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    mutate(perfstat = case_when(
      perfstat == "firstyearrebuildtoblim" ~ "precblim",
      perfstat == "firstyearrebuildtobpa"  ~ "precbpa",
      TRUE                                 ~ perfstat
    ))
  
  # title
  t <- paste(
    toupper(mystock),
    paste(unique(myassess), collapse="-"),
    paste(unique(myassessyear), collapse="-"),
    paste(unique(mymethod), collapse="-"),
    paste(unique(myom), collapse="-"),
    paste(unique(mymp), collapse="-"),
    sep="_"
  )
  
  plist <- list(NULL)
  i <- 1
  for (i in 1:length(myperfstat)) {
    if(myperfstat[[i]] %in% c("stock","fbar","catch","rec","tac")) {
      
      print(i)
      
      plist[[i]] <-
        
        ggplot(data=filter(d, perfstat==myperfstat[[i]]), 
               aes(x=year, y=mean, group=perfstat)) +
        theme_publication() +
        theme(plot.margin = unit(c(0,0,0,0), "cm")) +
        theme(legend.position="none") +
        theme( panel.grid.major.x = element_blank()) +
        {if (i == length(myperfstat)) 
          theme(axis.text.x = element_text(angle = 90, vjust=0.5)) else 
            theme(axis.text.x = element_blank()) } +
        
        # historical ribbon
        geom_ribbon(data=filter(h, !is.na(mean), perfstat==myperfstat[[i]]),
                    aes(ymin=lower, ymax=upper, group=iter), alpha=0.4, fill="gray") +
        
        # historical worms
        geom_line(data=filter(h, metric=="worm", perfstat==myperfstat[[i]]),
                  aes(x=year, y=data, group=iter), colour="darkgray", size=0.5) +
        
        # historical median
        geom_line(data=filter(h, !is.na(median), perfstat==myperfstat[[i]]), 
                  aes(x=year,y=median, group=iter), colour="black", size=1) +
    
        # future ribbon
        geom_ribbon(data=filter(d, !is.na(median), perfstat==myperfstat[[i]]),
                    aes(ymin=lower, ymax=upper, group=iter), alpha=0.4, fill=mycolours[[i]]) +
        
        # future worms
        geom_line(data=filter(d, metric=="worm", perfstat==myperfstat[[i]]),
                  aes(x=year, y=data, group=iter), colour=mycolours[[i]], size=0.5) +
        
        # future median
        geom_line(data=filter(d, !is.na(median), perfstat==myperfstat[[i]]),
                  aes(x=year,y=median,group=iter), colour=mycolours[[i]], size=1) +
        
        # vertical lines (periods)
        geom_vline(data=filter(p, period=="CU"),
                   aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="ST"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="MT"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="LT"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="LT"),
        #            aes(xintercept=maxyear), linetype="dotted") +
    
        # horizontal lines (reference points)
        {if (myperfstat[[i]] == "stock") geom_hline(aes(yintercept=blim), linetype="dashed") } +
        {if (myperfstat[[i]] == "stock") geom_text(aes(y=blim), label="blim", x=myfirstyear,
                                                   vjust=0, hjust=0) } +
        
        {if (myperfstat[[i]] == "stock") geom_hline(aes(yintercept=bpa), linetype="dotted") } +
        {if (myperfstat[[i]] == "stock") geom_text(aes(y=bpa), label="bpa", x=myfirstyear, 
                                                   vjust=0, hjust=0) } +
        
        {if (myperfstat[[i]] == "fbar") geom_hline(aes(yintercept=as.numeric(ftgt)), linetype="dashed") } +
        {if (myperfstat[[i]] == "fbar") geom_text(aes(y=as.numeric(ftgt)), label="ftgt", x=myfirstyear, 
                                                   vjust=0, hjust=0) } +

        expand_limits(y=0) +
        
        labs(x="", y = filter(u, perfstat == myperfstat[[i]])$unit2) +
        {if(i==1 & showtitle==TRUE) labs(title=t)} +
        
        facet_grid( perfstat ~ .)
      
  
    } else if (myperfstat[i] %in% c("pblim")) {
      
      print(i)
           
      plist[[i]] <-
        ggplot(data=filter(d, perfstat==myperfstat[[i]]), 
                   aes(x=year, y=mean, group=perfstat)) +
        theme_publication() +
        theme(plot.margin = unit(c(0,0,0,0), "cm")) +
        theme(legend.position="none") +
        theme( panel.grid.major.x = element_blank()) +
        {if (i == length(myperfstat)) theme(axis.text.x = element_text(angle = 90, vjust=0.5)) else theme(axis.text.x = element_blank()) } +
        
        geom_hline(yintercept=0.05, colour="gray", linetype="dashed") +
        
        geom_line(data=filter(h, perfstat == myperfstat[[i]]), colour="gray") +
        geom_line(colour=mycolours[[i]]) +
        
        # vertical lines (periods)
        geom_vline(data=filter(p, period=="CU"),
                   aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="ST"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="MT"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="LT"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="LT"),
        #            aes(xintercept=maxyear), linetype="dotted") +
    
        expand_limits(y=0) +
        
        labs(x="", y = filter(u, perfstat == myperfstat[[i]])$unit2) +
        {if(i==1 & showtitle==TRUE) labs(title=t)} +
        
        facet_grid( perfstat ~ .)
  
    } else if (myperfstat[i] %in% c("precbpa")) {
      
      print(i)
           
      plist[[i]] <-
        ggplot(data=filter(d, perfstat==myperfstat[[i]]), 
                   aes(x=year, y=mean, group=perfstat)) +
        theme_publication() +
        theme(plot.margin = unit(c(0,0,0,0), "cm")) +
        theme(legend.position="none") +
        theme( panel.grid.major.x = element_blank()) +
        {if (i == length(myperfstat)) theme(axis.text.x = element_text(angle = 90, vjust=0.5)) else theme(axis.text.x = element_blank()) } +
        
        geom_vline(data=p2, aes(xintercept=mean), colour=mycolours[[i]], linetype="dashed") +
        
        geom_hline(yintercept=0.5, colour="gray", linetype="dashed") +
        
        geom_vline(data=filter(p, period=="CU"),
                   aes(xintercept=minyear), linetype="dotted") +
        
        geom_text(data=filter(p2, perfstat %in% c("precbpa")),
                  aes(x=mean, label=ac(mean)),
                  y=0.0, vjust=0, hjust=0, nudge_x = 1, colour=mycolours[[i]]) +
    
        geom_line(colour=mycolours[[i]]) +
    
        expand_limits(y=0, x=min(p$minyear)) +
        
        labs(x="", y = filter(u, perfstat == myperfstat[[i]])$unit2) +
        {if(i==1 & showtitle==TRUE) labs(title=t)} +
        
        facet_grid( perfstat ~ .)

    } # end of if statements
  
  } # end of for loop

print(patchwork::wrap_plots(plist) + 
  patchwork::plot_layout(ncol = 1) +
  patchwork::plot_annotation() & theme(plot.tag = element_text(size = 8)))

ggsave(file = file.path(Res.dir,paste0(t, paste(myperfstat, collapse="_")," scenarioplot.png")),
       device="png", width = 30, height = 20, units = "cm")

  
} # end of function


# df %>% filter(perfstat=="stock") %>% View()

# plotscenario (
#     mystock      =  "WHOM",
#     myassess     = "SS3",
#     myassessyear = c("2020"),
#     mymethod     = "EQSIM",
#     myom         = c("OM2.3"),
#     mymp         = c("MP5.23"),
#     myftgt       = c(0.075),
#     myfacets     = c("perfstat"),
#     myfirstyear  = 2000,
#     mylastyear   = as.numeric(NA),
#     myperfstat   = c("stock","fbar"),
#     mycolours    = c("red","blue","red", "blue") )

# plotscenario (
#     mystock      =  "WHOM",
#     myassess     = "SS3",
#     myassessyear = c("2020"),
#     mymethod     = "EQSIM",
#     myom         = c("OM2.3"),
#     mymp         = c("MP5.23"),
#     myftgt       = c(0.074),
#     myfacets     = c("perfstat"),
#     myfirstyear  = 2000,
#     mylastyear   = as.numeric(NA),
#     myperfstat   = c("fbar"),
#     mycolours    = c("blue"),
#     showtitle    = FALSE)

plotscenario (
    mystock      =  "WHOM",
    myassess     = "SS3",
    myassessyear = c("2020"),
    mymethod     = "EQSIM",
    myom         = c("OM2.3"),
    mymp         = c("MP5.23"),
    myftgt       = c(0.074),
    myfacets     = c("perfstat"),
    myfirstyear  = 2000,
    mylastyear   = as.numeric(NA),
    myperfstat   = c("stock","fbar", "rec", "catch"),
    mycolours    = c("blue","red","green", "brown") )

plotscenario (
    mystock      =  "WHOM",
    myassess     = "SS3",
    myassessyear = c("2020"),
    mymethod     = "EQSIM",
    myom         = c("OM2.3"),
    mymp         = c("MP5.23"),
    myftgt       = c(0.074),
    myfacets     = c("perfstat"),
    myfirstyear  = 2000,
    mylastyear   = as.numeric(NA),
    myperfstat   = c("pblim","precbpa"),
    mycolours    = c("red","blue") )



  
```

*Working document for WKWHMRP 2021*


Martin Pastoors, `r format(Sys.time(), '%d/%m/%Y')`

&nbsp;  

**Abstract**

*To be done *  


### Constant F strategy

Simulation summary 2019 assessment, SS3, EqSim.

MP5.00 constant F; MP5.01 constant F with minimum TAC of 50kT; MP5.03 constant F with 20% IAV constraint above Btrigger. 

```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.00","MP5.01","MP5.03"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.00","MP5.01","MP5.03"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.00","MP5.01","MP5.03"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.00","MP5.01","MP5.03"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.00","MP5.01","MP5.03"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotrecovery (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.00","MP5.01","MP5.03"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

```
_fig x Simulation summary 2019 assessment, SS3, EqSim. MP5.00 constant F; MP5.01 constant F with minimum TAC of 50kT; MP5.03 constant F with 20% IAV constraint above Btrigger. _

##### page break

### ICES advice rule

Simulation summary 2019 assessment, SS3, EqSim.

MP5.10 ICES AR; MP5.11 ICES AR with minimum TAC of 50kT; MP5.13 ICES AR with 20% IAV constraint above Btrigger. 

```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotrecovery (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

```

_fig x Simulation summary 2019 assessment, SS3, EqSim. MP5.10 ICES AR; MP5.11 ICES AR with minimum TAC of 50kT; MP5.13 ICES AR with 20% IAV constraint above Btrigger._

##### page break

### Double breakpoint rule

Simulation summary 2019 assessment, SS3, EqSim.

MP5.20 Double BP; MP5.11 Double BP with minimum TAC of 50kT; MP5.13 Doubl BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy. 

```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotrecovery (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

```

_fig x Simulation summary 2019 assessment, SS3, EqSim. MP5.20 Double BP; MP5.21 Double BP with minimum TAC of 50kT; MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy._

### Double breakpoint rule

Simulation summary 2020 assessment, SS3, EqSim.

MP5.20 Double BP; MP5.11 Double BP with minimum TAC of 50kT; MP5.13 Doubl BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy. 

```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotrecovery (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

```

_fig x Simulation summary 2019 assessment, SS3, EqSim. MP5.20 Double BP; MP5.21 Double BP with minimum TAC of 50kT; MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy._

## SAM assessments in EqSim simulator

### ICES advice rule

Simulation summary 2019 assessment, SAM, EqSim.

MP5.10 ICES AR; MP5.11 ICES AR with minimum TAC of 50kT; MP5.13 ICES AR with 20% IAV constraint above Btrigger. 

```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotrecovery (
        mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.10","MP5.11","MP5.13"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

```

_fig x Simulation summary 2019 assessment, SAM, EqSim. MP5.10 ICES AR; MP5.11 ICES AR with minimum TAC of 50kT; MP5.13 ICES AR with 20% IAV constraint above Btrigger._

##### page break

### Double breakpoint rule

Simulation summary 2019 assessment, SAM, EqSim.

MP5.20 Double BP; MP5.21 Double BP with minimum TAC of 50kT; MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy. 

```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)


plotrecovery (
        mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4"),
        mymp         = c("MP5.20","MP5.21","MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

```

_fig x Simulation summary 2019 assessment, SAM, EqSim. MP5.20 Double BP; MP5.21 Double BP with minimum TAC of 50kT; MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy._

## SAM and SAM HCR forecast

SAM Assessment and SAM HCR forecast.  

MP5.10 ICES Advice Rule; MP5.20 Double BP; No IAV of fixed catch options. Minimum F in Double BP is 20% of Fmsy. 

```{r echo=FALSE, fig.align="center", fig.asp=0.5, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("SAMHCR"),
        myom         = c("no_OM"),
        mymp         = c("MP5.10","MP5.20"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("SAMHCR"),
        myom         = c("no_OM"),
        mymp         = c("MP5.10","MP5.20"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("SAMHCR"),
        myom         = c("no_OM"),
        mymp         = c("MP5.10","MP5.20"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("SAMHCR"),
        myom         = c("no_OM"),
        mymp         = c("MP5.10","MP5.20"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("SAMHCR"),
        myom         = c("no_OM"),
        mymp         = c("MP5.10","MP5.20"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)


plotrecovery (
        mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019"),
        mymethod     = c("SAMHCR"),
        myom         = c("no_OM"),
        mymp         = c("MP5.10","MP5.20"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("mp","ftgt"),
        myfirstyear  = 2000)

```

_fig x Simulation summary of SAM Assessment and SAM HCR forecast. MP5.10 ICES Advice Rule; MP5.20 Double BP; No IAV of fixed catch options. Minimum F in Double BP is 20% of Fmsy._

## Sensitivity analysis to assessment year as basis for the simulation

### Comparison of 2019 and 2020 SS3 assessment and EqSim simulator

MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy. 

SS3 assessment of 2019 and 2020 using  Eqsim simulator

```{r echo=FALSE, fig.align="center", fig.asp=0.5, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.3"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.3"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.3"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.3"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.3"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotrecovery (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.3"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

```
_fig x Simulation summary of SS3 Assessment 2019 and 2020 and EqSim simulator. MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy._

### Comparison of 2019 and 2020 SS3 assessment and EqSim simulator

MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy. 

SAM assessment of 2019 and 2020 using  Eqsim simulator

```{r echo=FALSE, fig.align="center", fig.asp=0.5, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4","OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4","OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4","OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4","OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4","OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

plotrecovery (
        mystock      = "WHOM",
        myassess     = "SAM",
        myassessyear = c("2019","2020"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.4","OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("assessyear","ftgt"),
        myfirstyear  = 2000)

```

_fig x Simulation summary of SAM Assessment 2019 and 2020 and EqSim simulator. MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy._


### Sensitivity of 2019 SS3 with different recruitment assumptions

2019 SS3 assessment using Eqsim simulator and different recruitment assumptions. MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy. 

Recruitment assumptions:

1. base case: recruitment from 2019 SS3 assessments. The geo mean of the full time series (ex 1982) is 2,572,649. The mean of the last 5 years from the 2019 assessment is 3,612,642

2. RR.V5: 2014-2018 recruitments reduced to the GM of the time series from 2002-2013. Apart from 2008, this represents an extended period of low recruitment. The geo mean for 2002-2013 is 1,620,516

3. RR.V6: 2014-2018 recruitments reduced to ½ of the baseline estimate. The mean of the last 5 years is then 1 806 321, similar to scenario RR.V5. 

4. RR.V7: 2014-2018 recruitments reduced to the mean of the 5 lowest from the full time series. The mean of the 5 lowest is 1,001,051. This is the most pessimistic scenario.


```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.2.RR.V5", "OM2.2.RR.V6", "OM2.2.RR.V7"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "stock",
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("om","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.2.RR.V5", "OM2.2.RR.V6", "OM2.2.RR.V7"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "fbar",
        mycolour     = "red",
        myyintercept = "fmsy",
        myvalue      = "median",
        myfacets     = c("om","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.2.RR.V5", "OM2.2.RR.V6", "OM2.2.RR.V7"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "rec",
        mycolour     = "darkgreen",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("om","ftgt"),
        myfirstyear  = 2000)

plotvar(mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.2.RR.V5", "OM2.2.RR.V6", "OM2.2.RR.V7"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myperfstat   = "catch",
        mycolour     = "brown",
        myyintercept = ac(NA),
        myvalue      = "median",
        myfacets     = c("om","ftgt"),
        myfirstyear  = 2000)

plotpblim (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.2.RR.V5", "OM2.2.RR.V6", "OM2.2.RR.V7"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myfacets     = c("om","ftgt"),
        myfirstyear  = 2000)

plotrecovery (
        mystock      = "WHOM",
        myassess     = "SS3",
        myassessyear = c("2019"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.2","OM2.2.RR.V5", "OM2.2.RR.V6", "OM2.2.RR.V7"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.0, 0.05, 0.075, 0.10, 0.15),
        myyintercept = 0.5,
        myfacets     = c("om","ftgt"),
        myfirstyear  = 2000)

# t <-
#   df %>% 
#   filter(perfstat %in% c("firstyearrebuildtoblim","firstyearrebuildtobpa")) %>% 
#   filter(mp %in% c("MP5.23")) %>% 
#   filter(assess %in% c("SS3")) 

```

_fig x Simulation summary of SS3 Assessment 2019, EqSim simulator and different recruitment assumptions. MP5.23 Double BP with 20% IAV constraint above Btrigger. Minimum F in Double BP is 20% of Fmsy._


### First year rebuild to a biomass reference point under different recruitment assumptions

```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

t <-
  df %>% 
  filter(mp %in% c("MP5.23")) %>% 
  filter(!grepl("WR", om)) %>% 
  filter(assess %in% c("SS3", "SAM")) %>% 
  filter(!grepl("RR$|lowest$",om)) %>% 
  filter(perfstat %in% c("precblim","precbpa")) %>%
  # mutate(year = as.numeric(year) - as.numeric(assessyear)) %>% 
  mutate(recscenario = gsub("OM2\\.2|OM2\\.3|OM2\\.4|OM2\\.5","", om)) %>% 
  mutate(recscenario = gsub("^\\.","", recscenario)) %>% 
  mutate(recscenario = ifelse(recscenario=="", "def", recscenario)) %>% 
  mutate(scenario    = paste(assess, assessyear)) %>% 
  group_by(assessyear, scenario, recscenario, mp, perfstat, ftgt) %>% 
  filter(mean >= rebuiltThreshold) %>% 
  group_by(assessyear, method, scenario, recscenario, mp, perfstat, ftgt) %>%
  summarise(year = min(year)) 

t1 <-
  t %>% 
  filter(ftgt == 0.0) %>% 
  dplyr::select(ftgt, year, perfstat, scenario, recscenario)

t2 <-
  t %>% 
  group_by(scenario, recscenario, perfstat) %>% 
  filter(ftgt == max(ftgt, na.rm=TRUE)) %>% 
  dplyr::select(ftgt, year, perfstat, scenario, recscenario)

t3 <-
  t %>% 
  filter(ftgt == 0.075) %>% 
  group_by(scenario, recscenario, perfstat) %>% 
  filter(ftgt == max(ftgt, na.rm=TRUE)) %>% 
  dplyr::select(ftgt, year, perfstat, scenario, recscenario)

t %>% 
  reshape2::dcast(perfstat+scenario+recscenario ~ ftgt, value.var="year", sum, fill=Inf, margins=FALSE) %>% 
  group_by(perfstat) %>%
  do(add_row(., .after=0)) %>%
    
  ungroup() %>% 
  pander::pandoc.table(., 
               style        = "simple",
               split.tables = 100, 
               split.cells  = c(rep(7,10)),
               justify      = "right",
               missing      = " ",
               big.mark     = '', 
               round        = c(0))


t %>% 
  group_by(scenario) %>% 
  mutate(id = group_indices()) %>% 
  mutate(offset = id * 0.1) %>% 
  mutate(year = year + offset) %>% 
  
  ggplot(aes(x=ftgt, y=year, group=scenario)) +
  theme_publication() +
  geom_point(aes(colour=scenario)) +
  geom_line(aes(colour=scenario)) +
  ggrepel::geom_text_repel(data=t1, aes(x=ftgt, y=year, label=ac(year), colour=scenario), 
                           force_pull = 0.5, nudge_x = -0.5,direction = "y", hjust = "right") +
  ggrepel::geom_text_repel(data=t2, aes(x=ftgt, y=year, label=ac(year), colour=scenario), 
                           force_pull = 0.5, nudge_x = 0.5,direction = "y", hjust = "left") +
  ggrepel::geom_text_repel(data=t3, aes(x=ftgt, y=year, label=ac(year), colour=scenario), 
                           force_pull = 0.5, nudge_x = 0.0,direction = "y", hjust = "middle") +
  # scale_y_continuous(limits=c(2020,2040),breaks=seq(2020,2040,5)) +
  facet_grid(perfstat~recscenario)


```

### First year rebuild to a biomass reference point under different reference point assumptions

```{r echo=FALSE, fig.align="center", fig.asp=0.65, message=FALSE, warning=FALSE}

t <-
  df %>% 
  filter(mp %in% c("MP5.23")) %>% 
  filter(!grepl("RR", om)) %>% 
  filter(assess %in% c("SS3", "SAM")) %>% 
  filter(perfstat %in% c("precblim","precbpa")) %>%
  # mutate(year = as.numeric(year) - as.numeric(assessyear)) %>% 
  mutate(rpscenario = gsub("OM2\\.2|OM2\\.3|OM2\\.4|OM2\\.5","", om)) %>% 
  mutate(rpscenario = gsub("^\\.","", rpscenario)) %>% 
  mutate(rpscenario = ifelse(rpscenario=="", "def", rpscenario)) %>% 
  mutate(scenario   = paste(assess, assessyear)) %>% 
  group_by(assessyear, scenario, rpscenario, mp, perfstat, ftgt) %>% 
  filter(mean >= rebuiltThreshold) %>% 
  group_by(assessyear, method, scenario, rpscenario, mp, perfstat, ftgt) %>%
  summarise(year = min(year)) %>% 
  mutate(assessment = stringr::word(scenario, 1, sep=" ")) %>% 
  mutate(assessmentyear = stringr::word(scenario, 2, sep=" ")) 


# t %>% ungroup() %>% distinct(scenario, rpscenario) %>% View()
# t %>% filter(ftgt==0.10, perfstat=="precbpa") %>% View()
t %>% 
  reshape2::dcast(perfstat+scenario+rpscenario ~ ftgt, value.var="year", sum, fill=Inf, margins=FALSE) %>% 
  group_by(perfstat) %>%
  do(add_row(., .after=0)) %>%
    
  ungroup() %>% 
  pander::pandoc.table(., 
               style        = "simple",
               split.tables = 100, 
               split.cells  = c(rep(7,10)),
               justify      = "right",
               missing      = " ",
               big.mark     = '', 
               round        = c(0))


t1 <-
  t %>% 
  filter(ftgt == 0.0) %>% 
  filter(perfstat == "precbpa") %>% 
  dplyr::select(ftgt, year, perfstat, scenario, rpscenario, assessment, assessmentyear)

t2 <-
  t %>% 
  group_by(scenario, rpscenario, perfstat) %>% 
  filter(ftgt == max(ftgt, na.rm=TRUE)) %>% 
  filter(perfstat == "precbpa") %>% 
  dplyr::select(ftgt, year, perfstat, scenario, rpscenario, assessment, assessmentyear)

t3 <-
  t %>% 
  filter(ftgt == 0.075) %>% 
  group_by(scenario, rpscenario, perfstat) %>% 
  filter(ftgt == max(ftgt, na.rm=TRUE)) %>% 
  filter(perfstat == "precbpa") %>% 
  dplyr::select(ftgt, year, perfstat, scenario, rpscenario, assessment, assessmentyear)

# t %>% 
#   group_by(scenario) %>% 
#   mutate(id = group_indices()) %>% 
#   mutate(offset = id * 0.1) %>% 
#   mutate(year = year + offset) %>% 
#   
#   ggplot(aes(x=ftgt, y=year, group=scenario)) +
#   theme_publication() +
#   geom_point(aes(colour=scenario)) +
#   geom_line(aes(colour=scenario)) +
#   ggrepel::geom_text_repel(data=t1, aes(x=ftgt, y=year, label=ac(year), colour=scenario), 
#                            force_pull = 0.5, nudge_x = -0.5,direction = "y", hjust = "right") +
#   ggrepel::geom_text_repel(data=t2, aes(x=ftgt, y=year, label=ac(year), colour=scenario), 
#                            force_pull = 0.5, nudge_x = 0.5,direction = "y", hjust = "left") +
#   ggrepel::geom_text_repel(data=t3, aes(x=ftgt, y=year, label=ac(year), colour=scenario), 
#                            force_pull = 0.5, nudge_x = 0.0,direction = "y", hjust = "middle") +
#   # scale_y_continuous(limits=c(2020,2040),breaks=seq(2020,2040,5)) +
#   facet_grid(perfstat~rpscenario)


t %>% 
  group_by(scenario) %>% 
  mutate(id = group_indices()) %>% 
  mutate(offset = id * 0.1) %>% 
  mutate(year = year + offset) %>% 
  filter(perfstat=="precbpa") %>% 
  
  ggplot(aes(x=ftgt, y=year, group=scenario)) +
  theme_publication() +
  geom_point(aes(colour=assessmentyear)) +
  geom_line(aes(colour=assessmentyear)) +
  ggrepel::geom_text_repel(data=t1, aes(x=ftgt, y=year, label=ac(year), colour=assessmentyear), 
                           force_pull = 0.5, nudge_x = -0.5,direction = "y", hjust = "right") +
  ggrepel::geom_text_repel(data=t2, aes(x=ftgt, y=year, label=ac(year), colour=assessmentyear), 
                           force_pull = 0.5, nudge_x = 0.5,direction = "y", hjust = "left") +
  ggrepel::geom_text_repel(data=t3, aes(x=ftgt, y=year, label=ac(year), colour=assessmentyear), 
                           force_pull = 0.5, nudge_x = 0.0,direction = "y", hjust = "middle") +
  # scale_y_continuous(limits=c(2020,2040),breaks=seq(2020,2040,5)) +
  facet_grid(assessment~rpscenario)

```

```{r echo=FALSE, fig.align="center", fig.asp=1.3, message=FALSE, warning=FALSE}

plotscenario (
    mystock      =  "WHOM",
    myassess     = "SS3",
    myassessyear = c("2020"),
    mymethod     = "EQSIM",
    myom         = c("OM2.3"),
    mymp         = c("MP5.23"),
    myftgt       = c(0.075),
    myfacets     = c("perfstat"),
    myfirstyear  = 2000,
    mylastyear   = as.numeric(NA),
    myperfstat   = c("stock","fbar", "rec", "catch"),
    mycolours    = c("blue","red","green", "brown") )


```


```{r echo=FALSE, fig.align="center", fig.asp=1.3, message=FALSE, warning=FALSE}

plotscenario (
    mystock      =  "WHOM",
    myassess     = "SS3",
    myassessyear = c("2020"),
    mymethod     = "EQSIM",
    myom         = c("OM2.3"),
    mymp         = c("MP5.23"),
    myftgt       = c(0.075),
    myfacets     = c("perfstat"),
    myfirstyear  = 2000,
    mylastyear   = as.numeric(NA),
    myperfstat   = c("pblim","precbpa"),
    mycolours    = c("red","blue") )


```
